All statics are now forbidden in lambda classes now.
Previously, we explicitly disallowed static *methods* in lambda
classes. Now, anything for which is_static() returns true is
disallowed. This includes not just static methods but also constant
declarations. Theoretically, we could specifically allow constants in
a lambda class, provided they do not close around anything and use the
outermost symbol table. But frankly, that's a lot of complexity, and I
really struggle to see a use case for it, so I'm disallowing it in the
spirit of simplicity.