Util package transferred over fully
Note that we omit fold1, which is equivalent to Rust's
Iterator::reduce. We also omit the `one` module, equivalent to
std::iter::once. For the former, I think I just wasn't aware of the
function when I wrote my version. For the latter, I think
std::iter::once was nightly when I started this project but is stable now